requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
distro==1.9.0
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.13.2
pydantic-core==2.46.2
python-dotenv==1.2.2
sniffio==1.3.1
typing-extensions==4.15.0
typing-inspection==0.4.2
